Create Bot 1.6.1
This version of the code moves the mouse in a random fashion either to the right or to the left, depending on the value of the direction parameter passed to the move function. It also removes the time.sleep calls in the share function, as they were not necessary. Additionally, it adds a check to ensure that the x and y variables returned by locateCenterOnScreen are not None before attempting to click on them, as locateCenterOnScreen returns None if it is unable to locate the image. Program ending error from unable to process None type is corrected.